requirements.txt
Django
python-dotenv
djangorestframework
setuptools
mysqlclient
whitenoise
gunicorn